Re-arranged dialog classes in a dialogs package. - Each dialog now has its own source file. (Except for GlobalHotkeyDialog, which is in hotkeysettings.py together with its base class.) - Package __init__.py exports all dialog classes. (This is the moved and renamed dialogs.py file.)
