+++ lib/autokey/qtui/dialogs/detectdialog.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from .common import inherits_from_ui_file_with_name
+from ..common import inherits_from_ui_file_with_name
 
 
 class DetectDialog(*inherits_from_ui_file_with_name("detectdialog")):
