+++ lib/autokey/qtui/dialogs/recorddialog.py
@@ -14,7 +14,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from .common import inherits_from_ui_file_with_name
+from ..common import inherits_from_ui_file_with_name
+
 import logging
 logger = logging.getLogger("root").getChild("Qt-GUI").getChild("Record Dialog")  # type: logging.Logger
 
@@ -48,4 +49,3 @@ class RecordDialog(*inherits_from_ui_file_with_name("record_dialog")):
             self.get_record_keyboard(), self.get_record_mouse(), self.get_delay()
         ))
         self.closure(False, self.get_record_keyboard(), self.get_record_mouse(), self.get_delay())
-
